+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.29.0] - 2026-07-04
+
 ### Security
 
 - **Nur-Lese-Konten werden korrekt als „Leser" geführt und serverseitig